requirements.txt
altgraph==0.17.4
attrs==25.1.0
certifi==2024.12.14
cffi==1.17.1
colorama==0.4.6
h11==0.14.0
idna==3.10
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.2
pefile==2023.2.7
pluggy==1.5.0
pycparser==2.22
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
PySocks==1.7.1
pytest==8.3.4
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
selenium==4.28.1
setuptools==75.8.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
